restructure the unit-tests
