+++ test/unit-test/CMakeLists.txt
@@ -84,8 +84,10 @@ add_cmock_targets()
 include( ${MODULE_ROOT_DIR}/test/unit-test/cmock/create_test.cmake )
 
 # Include unit-test build configuration.
-include( ${UNIT_TEST_DIR}/packetization/ut.cmake )
-include( ${UNIT_TEST_DIR}/depacketization/ut.cmake )
+include( ${UNIT_TEST_DIR}/g711/ut.cmake )
+include( ${UNIT_TEST_DIR}/opus/ut.cmake )
+include( ${UNIT_TEST_DIR}/h264/ut.cmake)
+include( ${UNIT_TEST_DIR}/vp8/ut.cmake)
 include( ${UNIT_TEST_DIR}/rtp_packet_queue/ut.cmake )
 
 #  ==================================== Coverage Analysis configuration ========================================
@@ -94,8 +96,10 @@ add_custom_target( coverage
     COMMAND ${CMAKE_COMMAND} -DCMOCK_DIR=${CMOCK_DIR}
     -P ${MODULE_ROOT_DIR}/test/unit-test/cmock/coverage.cmake
     DEPENDS cmock unity
-    packetization
-    depacketization
+    g711
+    opus
+    h264
+    vp8
     rtp_packet_queue
     WORKING_DIRECTORY ${CMAKE_BINARY_DIR}
 )
