+++ test/unit-test/vp8/ut.cmake
@@ -3,7 +3,7 @@
 include( ${MODULE_ROOT_DIR}/rtpFilePaths.cmake )
 
 # ====================  Define your project name (edit) ========================
-set( project_name "depacketization" )
+set( project_name "vp8" )
 
 message( STATUS "${project_name}" )
 
@@ -11,9 +11,6 @@ message( STATUS "${project_name}" )
 
 # List the files to mock here.
 list(APPEND mock_list
-            "${MODULE_ROOT_DIR}/codec_packetizers/g711/include/g711_data_types.h"
-            "${MODULE_ROOT_DIR}/codec_packetizers/h264/include/h264_data_types.h"
-            "${MODULE_ROOT_DIR}/codec_packetizers/opus/include/opus_data_types.h"
             "${MODULE_ROOT_DIR}/codec_packetizers/vp8/include/vp8_data_types.h"
         )
 # List the directories your mocks need.
@@ -31,10 +28,8 @@ list(APPEND mock_define_list
 
 # List the files you would like to test here.
 list(APPEND real_source_files
-            ${MODULE_ROOT_DIR}/codec_packetizers/g711/g711_depacketizer.c
-            ${MODULE_ROOT_DIR}/codec_packetizers/h264/h264_depacketizer.c
-            ${MODULE_ROOT_DIR}/codec_packetizers/opus/opus_depacketizer.c
             ${MODULE_ROOT_DIR}/codec_packetizers/vp8/vp8_depacketizer.c
+            ${MODULE_ROOT_DIR}/codec_packetizers/vp8/vp8_packetizer.c
         )
 # List the directories the module under test includes.
 list(APPEND real_include_directories
