+++ test/unit-test/vp8/vp8_utest.c
@@ -7,131 +7,38 @@
 #include <stdint.h>
 
 /* API includes. */
-#include "h264_packetizer.h"
-#include "g711_packetizer.h"
 #include "vp8_packetizer.h"
-#include "opus_packetizer.h"
+#include "vp8_depacketizer.h"
 
 /* ===========================  EXTERN VARIABLES  =========================== */
 
-#define MAX_NALUS_IN_A_FRAME        512
-#define MAX_H264_PACKET_LENGTH      12
+#define MAX_FRAME_LENGTH        10 * 1024
+#define MAX_PACKET_IN_A_FRAME   512
+
 #define PACKETIZATION_BUFFER_LENGTH 10
 
+#define VP8_PACKETS_ARR_LEN     10
+#define VP8_FRAME_BUF_LEN       32
+
 uint8_t packetizationBuffer[ PACKETIZATION_BUFFER_LENGTH];
+uint8_t frameBuffer[ MAX_FRAME_LENGTH ];
 
 void setUp( void )
 {
     memset( &( packetizationBuffer[ 0 ] ),
             0,
             sizeof( packetizationBuffer ) );
+    memset( &( frameBuffer[ 0 ] ),
+            0,
+            sizeof( frameBuffer ) );
 }
 
 void tearDown( void )
 {
 }
 
-/* ==============================  Test Cases  ============================== */
 
-/**
- * @brief Validate Opus packetization happy path.
- */
-void test_Opus_Packetizer( void )
-{
-    OpusResult_t result;
-    uint8_t frameData[] = { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05,
-                            0x10, 0x11, 0x12, 0x13, 0x14, 0x15,
-                            0x20, 0x21, 0x22, 0x23, 0x24, 0x25,
-                            0x30, 0x31, 0x32, 0x33, 0x34, 0x35,
-                            0x40, 0x41 };
-    OpusPacketizerContext_t ctx;
-    OpusPacket_t pkt;
-    OpusFrame_t frame;
-    size_t curFrameDataIndex = 0;
-
-    frame.pFrameData = &( frameData [ 0 ] );
-    frame.frameDataLength = sizeof( frameData );
-
-    result = OpusPacketizer_Init( &( ctx ),
-                                  &( frame ) );
-    TEST_ASSERT_EQUAL( result,
-                       OPUS_RESULT_OK );
-
-    pkt.pPacketData = &( packetizationBuffer[ 0 ] );
-    pkt.packetDataLength = PACKETIZATION_BUFFER_LENGTH;
-    result = OpusPacketizer_GetPacket( &( ctx ),
-                                       &( pkt ) );
-
-    while( result != OPUS_RESULT_NO_MORE_PACKETS )
-    {
-        TEST_ASSERT_TRUE( pkt.packetDataLength <= PACKETIZATION_BUFFER_LENGTH );
-
-        TEST_ASSERT_EQUAL_UINT8_ARRAY( &( frameData[ curFrameDataIndex ] ),
-                                       &( pkt.pPacketData[ 0 ] ),
-                                       pkt.packetDataLength );
-        curFrameDataIndex += pkt.packetDataLength;
-
-        pkt.pPacketData = &( packetizationBuffer[ 0 ] );
-        pkt.packetDataLength = PACKETIZATION_BUFFER_LENGTH;
-        result = OpusPacketizer_GetPacket( &( ctx ),
-                                           &( pkt ) );
-    }
-
-    TEST_ASSERT_EQUAL( curFrameDataIndex,
-                       sizeof( frameData ) );
-}
-
-/*-----------------------------------------------------------*/
-
-/**
- * @brief Validate G711 packetization happy path.
- */
-void test_G711_Packetizer( void )
-{
-    G711Result_t result;
-    uint8_t frameData[] = { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05,
-                            0x10, 0x11, 0x12, 0x13, 0x14, 0x15,
-                            0x20, 0x21, 0x22, 0x23, 0x24, 0x25,
-                            0x30, 0x31, 0x32, 0x33, 0x34, 0x35,
-                            0x40, 0x41 };
-    G711PacketizerContext_t ctx;
-    G711Packet_t pkt;
-    G711Frame_t frame;
-    size_t curFrameDataIndex = 0;
-
-    frame.pFrameData = &( frameData [ 0 ] );
-    frame.frameDataLength = sizeof( frameData );
-
-    result = G711Packetizer_Init( &( ctx ),
-                                  &( frame ) );
-    TEST_ASSERT_EQUAL( result,
-                       G711_RESULT_OK );
-
-    pkt.pPacketData = &( packetizationBuffer[ 0 ] );
-    pkt.packetDataLength = PACKETIZATION_BUFFER_LENGTH;
-    result = G711Packetizer_GetPacket( &( ctx ),
-                                       &( pkt ) );
-
-    while( result != G711_RESULT_NO_MORE_PACKETS )
-    {
-        TEST_ASSERT_TRUE( pkt.packetDataLength <= PACKETIZATION_BUFFER_LENGTH );
-
-        TEST_ASSERT_EQUAL_UINT8_ARRAY( &( frameData[ curFrameDataIndex ] ),
-                                       &( pkt.pPacketData[ 0 ] ),
-                                       pkt.packetDataLength );
-        curFrameDataIndex += pkt.packetDataLength;
-
-        pkt.pPacketData = &( packetizationBuffer[ 0 ] );
-        pkt.packetDataLength = PACKETIZATION_BUFFER_LENGTH;
-        result = G711Packetizer_GetPacket( &( ctx ),
-                                           &( pkt ) );
-    }
-
-    TEST_ASSERT_EQUAL( curFrameDataIndex,
-                       sizeof( frameData ) );
-}
-
-/*-----------------------------------------------------------*/
+/* ==============================  Test Cases for Packetization ============================== */
 
 /**
  * @brief Validate VP8 packetization happy path.
@@ -144,7 +51,7 @@ void test_VP8_Packetizer( void )
                             0x20, 0x21, 0x22, 0x23, 0x24, 0x25,
                             0x30, 0x31, 0x32, 0x33, 0x34, 0x35,
                             0x40, 0x41 };
-    VP8PacketizerContext_t ctx;
+    VP8PacketizerContext_t ctx = { 0 };
     VP8Frame_t frame;
     VP8Packet_t pkt;
     uint8_t payloadDescFirstPkt[] = { 0x10 };
@@ -219,7 +126,7 @@ void test_VP8_Packetizer_PictureID( void )
                             0x20, 0x21, 0x22, 0x23, 0x24, 0x25,
                             0x30, 0x31, 0x32, 0x33, 0x34, 0x35,
                             0x40, 0x41 };
-    VP8PacketizerContext_t ctx;
+    VP8PacketizerContext_t ctx = { 0 };
     VP8Frame_t frame;
     VP8Packet_t pkt;
     uint8_t payloadDescFirstPkt[] = { 0xB0, 0x80, 0x1A };
@@ -297,7 +204,7 @@ void test_VP8_Packetizer_TID( void )
                             0x20, 0x21, 0x22, 0x23, 0x24, 0x25,
                             0x30, 0x31, 0x32, 0x33, 0x34, 0x35,
                             0x40, 0x41 };
-    VP8PacketizerContext_t ctx;
+    VP8PacketizerContext_t ctx = { 0 };
     VP8Frame_t frame;
     VP8Packet_t pkt;
     uint8_t payloadDescFirstPkt[] = { 0xB0, 0x20, 0x60 };
@@ -375,7 +282,7 @@ void test_VP8_Packetizer_TL0PICIDX( void )
                             0x20, 0x21, 0x22, 0x23, 0x24, 0x25,
                             0x30, 0x31, 0x32, 0x33, 0x34, 0x35,
                             0x40, 0x41 };
-    VP8PacketizerContext_t ctx;
+    VP8PacketizerContext_t ctx = { 0 };
     VP8Frame_t frame;
     VP8Packet_t pkt;
     uint8_t payloadDescFirstPkt[] = { 0xB0, 0x40, 0xAB };
@@ -453,7 +360,7 @@ void test_VP8_Packetizer_AllProperties( void )
                             0x20, 0x21, 0x22, 0x23, 0x24, 0x25,
                             0x30, 0x31, 0x32, 0x33, 0x34, 0x35,
                             0x40, 0x41 };
-    VP8PacketizerContext_t ctx;
+    VP8PacketizerContext_t ctx = { 0 };
     VP8Frame_t frame;
     VP8Packet_t pkt;
     uint8_t payloadDescFirstPkt[] = { 0xB0, 0xF0, 0xFA, 0xCD, 0xAB, 0xBA };
@@ -525,197 +432,199 @@ void test_VP8_Packetizer_AllProperties( void )
                        sizeof( frameData ) );
 }
 
-/*-----------------------------------------------------------*/
+
+/* ==============================  Test Cases for Depacketization ============================== */
+
+
 
 /**
- * @brief Validate H264 packetization happy path with adding Nalus.
+ * @brief Validate VP8 depacketization happy path.
  */
-void test_H264_Packetizer_AddNalu( void )
+
+void test_VP8_Depacketizer_AllProperties( void )
 {
-    uint8_t pFrame[] = { 0x00, 0x00, 0x00, 0x01, 0x09, 0x10,
-                         0x00, 0x00, 0x00, 0x01, 0x67, 0x42, 0xc0, 0x1f, 0xda,
-                                                 0x01, 0x40, 0x16, 0xec, 0x05,
-                                                 0xa8, 0x08,
-                         0x00, 0x00, 0x00, 0x01, 0x68, 0xce, 0x3c, 0x80,
-                         0x00, 0x00, 0x00, 0x01, 0x06, 0x05, 0xff, 0xff, 0xb7,
-                                                 0xdc, 0x45, 0xe9, 0xbd, 0xe6,
-                                                 0xd9, 0x48,
-                         0x00, 0x00, 0x00, 0x01, 0x65, 0x88, 0x84, 0x12, 0xff,
-                                                 0xff, 0xfc, 0x3d, 0x14, 0x00,
-                                                 0x04, 0xba, 0xeb, 0xae, 0xba,
-                                                 0xeb, 0xae, 0xba, 0xeb, 0xae,
-                                                 0xba,
-                         0x00, 0x00, 0x00, 0x01, 0x65, 0x00, 0x6e, 0x22, 0x21,
-                                                 0x04, 0xbf, 0xff, 0xff, 0x0f,
-                                                 0x45, 0x00 };
-    size_t frameLength = sizeof( pFrame );
-    size_t curIndex = 0, naluStartIndex = 0, remainingLength;
-    uint32_t fistStartCode = 1;
-    H264PacketizerContext_t ctx;
-    H264Result_t result;
-    H264Packet_t pkt;
-    uint8_t pktBuffer[ MAX_H264_PACKET_LENGTH ];
-    Nalu_t nalusArray[ MAX_NALUS_IN_A_FRAME ], nalu;
-    uint32_t packetNumber = 0;
-    uint32_t expectedPacketLength[] = { 2, 12, 4, 12, 12, 12, 12 };
+    VP8DepacketizerContext_t ctx = { 0 };
+    VP8Result_t result;
+    VP8Packet_t pkt;
+    VP8Frame_t frame;
+    VP8Packet_t packetsArray[ VP8_PACKETS_ARR_LEN ];
 
-    uint8_t startCode1[] = { 0x00, 0x00, 0x00, 0x01 };
-    uint8_t startCode2[] = { 0x00, 0x00, 0x01 };
+    uint8_t packetData1[] = { 0xB0, 0xF0, 0xFA, 0xCD, 0xAB, 0xBA, 0x00, 0x01, 0x02, 0x03 };
+    uint8_t packetData2[] = { 0xA0, 0xF0, 0xFA, 0xCD, 0xAB, 0xBA, 0x04, 0x05, 0x10, 0x11 };
+    uint8_t packetData3[] = { 0xA0, 0xF0, 0xFA, 0xCD, 0xAB, 0xBA, 0x12, 0x13, 0x14, 0x15 };
+    uint8_t packetData4[] = { 0xA0, 0xF0, 0xFA, 0xCD, 0xAB, 0xBA, 0x20, 0x21, 0x22, 0x23 };
+    uint8_t packetData5[] = { 0xA0, 0xF0, 0xFA, 0xCD, 0xAB, 0xBA, 0x24, 0x25, 0x30, 0x31 };
+    uint8_t packetData6[] = { 0xA0, 0xF0, 0xFA, 0xCD, 0xAB, 0xBA, 0x32, 0x33, 0x34, 0x35 };
+    uint8_t packetData7[] = { 0xA0, 0xF0, 0xFA, 0xCD, 0xAB, 0xBA, 0x40, 0x41 };
 
-    result = H264Packetizer_Init( &( ctx ),
-                                  &( nalusArray[ 0 ] ),
-                                  MAX_NALUS_IN_A_FRAME );
-    TEST_ASSERT_EQUAL( result,
-                       H264_RESULT_OK );
+    uint8_t decodedFrameData[] = { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05,
+                                   0x10, 0x11, 0x12, 0x13, 0x14, 0x15,
+                                   0x20, 0x21, 0x22, 0x23, 0x24, 0x25,
+                                   0x30, 0x31, 0x32, 0x33, 0x34, 0x35,
+                                   0x40, 0x41 };
 
-    while( curIndex < frameLength )
-    {
-        remainingLength = frameLength - curIndex;
+    result = VP8Depacketizer_Init( &( ctx ),
+                                   &( packetsArray[ 0 ] ),
+                                   VP8_PACKETS_ARR_LEN );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
-        /* Check the presence of 4 byte start code. */
-        if( remainingLength >= sizeof( startCode1 ) )
-        {
-            if( memcmp( &( pFrame[ curIndex ] ),
-                        &( startCode1[ 0 ] ),
-                        sizeof( startCode1 ) ) == 0 )
-            {
-                if( fistStartCode == 1 )
-                {
-                    fistStartCode = 0;
-                }
-                else
-                {
-                    nalu.pNaluData = &( pFrame[ naluStartIndex ] );
-                    nalu.naluDataLength = curIndex - naluStartIndex;
-                    result = H264Packetizer_AddNalu( &( ctx ),
-                                                     &( nalu ) );
-                    TEST_ASSERT_EQUAL( result,
-                                       H264_RESULT_OK );
-                }
+    pkt.pPacketData = &( packetData1[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData1 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
-                naluStartIndex = curIndex + sizeof( startCode1 );
-                curIndex = curIndex + sizeof( startCode1 );
-                continue;
-            }
-        }
+    pkt.pPacketData = &( packetData2[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData2 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
-        /* Check the presence of 3 byte start code. */
-        if( remainingLength >= sizeof( startCode2 ) )
-        {
-            if( memcmp( &( pFrame[ curIndex ] ),
-                        &( startCode2[ 0 ] ),
-                        sizeof( startCode2 ) ) == 0 )
-            {
-                if( fistStartCode == 1 )
-                {
-                    fistStartCode = 0;
-                }
-                else
-                {
-                    nalu.pNaluData = &( pFrame[ naluStartIndex ] );
-                    nalu.naluDataLength = curIndex - naluStartIndex;
-                    result = H264Packetizer_AddNalu( &( ctx ),
-                                                     &( nalu ) );
-                }
+    pkt.pPacketData = &( packetData3[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData3 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
-                naluStartIndex = curIndex + sizeof( startCode2 );
-                curIndex = curIndex + sizeof( startCode2 );
-                continue;
-            }
-        }
+    pkt.pPacketData = &( packetData4[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData4 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
-        curIndex += 1;
-    }
+    pkt.pPacketData = &( packetData5[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData5 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
-    if( naluStartIndex > 0 )
-    {
-        nalu.pNaluData = &( pFrame[ naluStartIndex ] );
-        nalu.naluDataLength = frameLength - naluStartIndex;
-        result = H264Packetizer_AddNalu( &( ctx ),
-                                         &( nalu ) );
-    }
+    pkt.pPacketData = &( packetData6[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData6 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
-    pkt.pPacketData = &( pktBuffer[ 0 ] );
-    pkt.packetDataLength = MAX_H264_PACKET_LENGTH;
-    result = H264Packetizer_GetPacket( &( ctx ),
-                                       &( pkt ) );
+    pkt.pPacketData = &( packetData7[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData7 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
+    frame.pFrameData = &( frameBuffer[ 0 ] );
+    frame.frameDataLength = VP8_FRAME_BUF_LEN;
+    result = VP8Depacketizer_GetFrame( &( ctx ),
+                                       &( frame ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
-    while( result != H264_RESULT_NO_MORE_PACKETS )
-    {
-        TEST_ASSERT_EQUAL( expectedPacketLength[ packetNumber ],
-                           pkt.packetDataLength );
-        packetNumber += 1;
+    TEST_ASSERT_EQUAL( sizeof( decodedFrameData ),
+                       frame.frameDataLength );
 
-        pkt.pPacketData = &( pktBuffer[ 0 ] );
-        pkt.packetDataLength = MAX_H264_PACKET_LENGTH;
-        result = H264Packetizer_GetPacket( &( ctx ),
-                                           &( pkt ) );
-    }
+    TEST_ASSERT_EQUAL_UINT8_ARRAY( &( decodedFrameData[ 0 ] ),
+                                   &( frame.pFrameData[ 0 ] ),
+                                   frame.frameDataLength );
+
+    TEST_ASSERT_EQUAL( 0x7ACD,
+                       frame.pictureId );
+    TEST_ASSERT_EQUAL( 0xAB,
+                       frame.tl0PicIndex );
+    TEST_ASSERT_EQUAL( 5,
+                       frame.tid );
+    TEST_ASSERT_EQUAL( 10,
+                       frame.keyIndex );
+    TEST_ASSERT_EQUAL( ( VP8_FRAME_PROP_NON_REF_FRAME |
+                         VP8_FRAME_PROP_PICTURE_ID_PRESENT |
+                         VP8_FRAME_PROP_TL0PICIDX_PRESENT |
+                         VP8_FRAME_PROP_TID_PRESENT |
+                         VP8_FRAME_PROP_KEYIDX_PRESENT |
+                         VP8_FRAME_PROP_DEPENDS_ON_BASE_ONLY ),
+                       frame.frameProperties );
 }
 
 /*-----------------------------------------------------------*/
 
 /**
- * @brief Validate H264 packetization happy path with adding Frame.
+ * @brief Validate VP8 depacketization happy path containing picture ID.
  */
-void test_H264_Packetizer_AddFrame( void )
+
+void test_VP8_Depacketizer_PictureID( void )
 {
-    uint8_t pFrame[] = { 0x00, 0x00, 0x00, 0x01, 0x09, 0x10,
-                         0x00, 0x00, 0x00, 0x01, 0x67, 0x42, 0xc0, 0x1f, 0xda,
-                                                 0x01, 0x40, 0x16, 0xec, 0x05,
-                                                 0xa8, 0x08,
-                         0x00, 0x00, 0x00, 0x01, 0x68, 0xce, 0x3c, 0x80,
-                         0x00, 0x00, 0x00, 0x01, 0x06, 0x05, 0xff, 0xff, 0xb7,
-                                                 0xdc, 0x45, 0xe9, 0xbd, 0xe6,
-                                                 0xd9, 0x48,
-                         0x00, 0x00, 0x00, 0x01, 0x65, 0x88, 0x84, 0x12, 0xff,
-                                                 0xff, 0xfc, 0x3d, 0x14, 0x00,
-                                                 0x04, 0xba, 0xeb, 0xae, 0xba,
-                                                 0xeb, 0xae, 0xba, 0xeb, 0xae,
-                                                 0xba,
-                         0x00, 0x00, 0x00, 0x01, 0x65, 0x00, 0x6e, 0x22, 0x21,
-                                                 0x04, 0xbf, 0xff, 0xff, 0x0f,
-                                                 0x45, 0x00 };
-    size_t frameLength = sizeof( pFrame );
-    H264PacketizerContext_t ctx;
-    H264Result_t result;
-    H264Packet_t pkt;
-    Frame_t frame;
-    Nalu_t nalusArray[ MAX_NALUS_IN_A_FRAME ];
-    uint8_t pktBuffer[ MAX_H264_PACKET_LENGTH ];
-    uint32_t packetNumber = 0;
-    uint32_t expectedPacketLength[] = { 2, 12, 4, 12, 12, 12, 12 };
+    VP8DepacketizerContext_t ctx = { 0 };
+    VP8Result_t result;
+    VP8Packet_t pkt;
+    VP8Frame_t frame;
+    VP8Packet_t packetsArray[ VP8_PACKETS_ARR_LEN ];
 
-    result = H264Packetizer_Init( &( ctx ),
-                                  &( nalusArray[ 0 ] ),
-                                  MAX_NALUS_IN_A_FRAME );
-    TEST_ASSERT_EQUAL( result,
-                       H264_RESULT_OK );
+    uint8_t packetData1[] = { 0xB0, 0x40, 0xAB, 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x10 };
+    uint8_t packetData2[] = { 0xA0, 0x40, 0xAB, 0x11, 0x12, 0x13, 0x14, 0x15, 0x20, 0x21 };
+    uint8_t packetData3[] = { 0xA0, 0x40, 0xAB, 0x22, 0x23, 0x24, 0x25, 0x30, 0x31, 0x32 };
+    uint8_t packetData4[] = { 0xA0, 0x40, 0xAB, 0x33, 0x34, 0x35, 0x40, 0x41 };
 
-    frame.pFrameData = pFrame;
-    frame.frameDataLength = frameLength;
-    result = H264Packetizer_AddFrame( &( ctx ),
-                                      &( frame ) );
-    TEST_ASSERT_EQUAL( result,
-                       H264_RESULT_OK );
+    uint8_t decodedFrameData[] = { 0x00, 0x01, 0x02, 0x03, 0x04, 0x05,
+                                   0x10, 0x11, 0x12, 0x13, 0x14, 0x15,
+                                   0x20, 0x21, 0x22, 0x23, 0x24, 0x25,
+                                   0x30, 0x31, 0x32, 0x33, 0x34, 0x35,
+                                   0x40, 0x41 };
 
-    pkt.pPacketData = &( pktBuffer[ 0 ] );
-    pkt.packetDataLength = MAX_H264_PACKET_LENGTH;
-    result = H264Packetizer_GetPacket( &( ctx ),
-                                       &( pkt ) );
+    result = VP8Depacketizer_Init( &( ctx ),
+                                   &( packetsArray[ 0 ] ),
+                                   VP8_PACKETS_ARR_LEN );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
-    while( result != H264_RESULT_NO_MORE_PACKETS )
-    {
-        TEST_ASSERT_EQUAL( expectedPacketLength[ packetNumber ],
-                           pkt.packetDataLength );
-        packetNumber += 1;
+    pkt.pPacketData = &( packetData1[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData1 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
 
-        pkt.pPacketData = &( pktBuffer[ 0 ] );
-        pkt.packetDataLength = MAX_H264_PACKET_LENGTH;
-        result = H264Packetizer_GetPacket( &( ctx ),
-                                           &( pkt ) );
-    }
+    pkt.pPacketData = &( packetData2[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData2 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
+
+    pkt.pPacketData = &( packetData3[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData3 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
+
+    pkt.pPacketData = &( packetData4[ 0 ] );
+    pkt.packetDataLength = sizeof( packetData4 );
+    result = VP8Depacketizer_AddPacket( &( ctx ),
+                                        &( pkt ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
+
+    frame.pFrameData = &( frameBuffer[ 0 ] );
+    frame.frameDataLength = VP8_FRAME_BUF_LEN;
+    result = VP8Depacketizer_GetFrame( &( ctx ),
+                                       &( frame ) );
+    TEST_ASSERT_EQUAL( VP8_RESULT_OK,
+                       result );
+
+    TEST_ASSERT_EQUAL( sizeof( decodedFrameData ),
+                       frame.frameDataLength );
+
+    TEST_ASSERT_EQUAL_UINT8_ARRAY( &( decodedFrameData[ 0 ] ),
+                                   &( frame.pFrameData[ 0 ] ),
+                                   frame.frameDataLength );
+
+    TEST_ASSERT_EQUAL( 0xAB,
+                       frame.tl0PicIndex );
+    TEST_ASSERT_EQUAL( ( VP8_FRAME_PROP_NON_REF_FRAME |
+                         VP8_FRAME_PROP_TL0PICIDX_PRESENT ),
+                       frame.frameProperties );
 }
-
-/*-----------------------------------------------------------*/
